requirements.txt
python-dotenv
gpiozero
fastapi
uvicorn[standard]
amcrest